refactor: 컨트롤러의 비즈니스 로직을 서비스로 이동
