+++ src/main/java/roomescape/repository/ReservationJDBCDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.repository;
 
 import java.sql.Date;
 import java.util.List;
