+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.repository;
 
 import java.util.List;
 import roomescape.entity.Reservation;
