+++ src/main/java/roomescape/service/ReservationService.java
@@ -3,19 +3,21 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.dao.ReservationRepository;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.entity.Reservation;
+import roomescape.repository.ReservationRepository;
 
 @Service
 @Transactional(readOnly = true)
 public class ReservationService {
 
     private final ReservationRepository reservationRepository;
+    private final ReservationTimeService timeService;
 
-    public ReservationService(ReservationRepository reservationRepository) {
+    public ReservationService(ReservationRepository reservationRepository, ReservationTimeService timeService) {
         this.reservationRepository = reservationRepository;
+        this.timeService = timeService;
     }
 
     public List<ReservationResponse> readReservation() {
@@ -26,6 +28,7 @@ public class ReservationService {
 
     @Transactional
     public ReservationResponse postReservation(ReservationRequest request) {
+        timeService.existsTimeById(request.timeId());
         Reservation newReservation = reservationRepository.save(request.toEntity(), request.timeId());
         return ReservationResponse.from(newReservation);
     }
