+++ src/test/java/roomescape/controller/ReservationControllerSpringExtensionTest.java
@@ -25,13 +25,13 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import roomescape.config.SpringExtensionTestConfig;
-import roomescape.dao.ReservationJDBCDao;
-import roomescape.dao.ReservationTimeJDBCDao;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.entity.Reservation;
 import roomescape.entity.ReservationTime;
 import roomescape.exceptions.EntityNotFoundException;
+import roomescape.repository.ReservationJDBCDao;
+import roomescape.repository.ReservationTimeJDBCDao;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
 
@@ -54,10 +54,9 @@ public class ReservationControllerSpringExtensionTest {
         MapSqlParameterSource params = new MapSqlParameterSource("startAt", "15:00");
 
         namedJdbcTemplate.update(timeSql, params, keyHolder, new String[]{"id"});
-
-        ReservationService service = new ReservationService(new ReservationJDBCDao(namedJdbcTemplate));
         ReservationTimeService timeService = new ReservationTimeService(new ReservationTimeJDBCDao(namedJdbcTemplate));
-        controller = new ReservationController(service, timeService);
+        ReservationService service = new ReservationService(new ReservationJDBCDao(namedJdbcTemplate), timeService);
+        controller = new ReservationController(service);
     }
 
     @Test
